refactor(stats): change type experiment number parameter

+++ src/stats.cpp
@@ -16,7 +16,7 @@ std::pair<double, double> confidence_interval(int percentage, double mean, doubl
     return std::make_pair(mean - offset, mean + offset);
 }
 
-void save_results(std::fstream &file, int N, stats &measurements) {
+void save_results(std::fstream &file, unsigned long long N, stats &measurements) {
     file << std::to_string(N) + " ";
     file << std::get<0>(measurements) << " ";
     file << std::get<1>(measurements) << " ";
